backend/requirements.txt
# Django Core
Django>=5.1,<5.2
djangorestframework>=3.15,<4.0

# Authentication
djangorestframework-simplejwt>=5.3,<6.0

# API Documentation
drf-spectacular>=0.27,<1.0

# Environment & Configuration
django-environ>=0.11,<1.0

# CORS
django-cors-headers>=4.3,<5.0

# Filtering
django-filter>=24.0,<25.0

# Production Server
gunicorn>=22.0,<23.0

# Static Files
whitenoise>=6.7,<7.0

# Database
psycopg[binary]>=3.1,<4.0

# Redis
redis>=5.0,<6.0
django-redis>=5.4,<6.0

# Image Processing
Pillow>=10.4,<12.0

# Utilities
requests>=2.31,<3.0

# Data Processing
pandas>=2.2,<3.0
openpyxl>=3.1,<4.0

